pyproject.toml
[project]
name = "apart-rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "django-rest-knox>=5.0.2",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "asgiref>=3.8.1",
    "sqlparse>=0.5.3",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.11.13",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "django-cors-headers>=4.7.0",
    "django-unfold>=0.60.0",
    "pillow>=11.2.1",
    "django-allauth>=65.10.0",
    "knox>=0.1.14",
    "whitenoise>=6.9.0",
    "setuptools>=80.9.0",
    "cryptography>=45.0.6",
    "whitenoise>=6.9.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I"]
ignore = []
fix = true
unsafe-fixes = false
show-fixes = true
preview = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
